Improved code organization.
